feat: integrate EngineArgs into NodeCommand
