+++ crates/node/core/src/args/mod.rs
@@ -56,5 +56,9 @@ pub use datadir_args::DatadirArgs;
 mod benchmark_args;
 pub use benchmark_args::BenchmarkArgs;
 
+/// EngineArgs for configuring the engine
+mod engine;
+pub use engine::EngineArgs;
+
 mod error;
 pub mod types;
